src/pipelines/cpu_pipeline.py
import os
import logging
import torch
from PIL import Image
from .base_pipeline import BasePipeline

logger = logging.getLogger(__name__)

class CPUPipeline(BasePipeline):
    """CPU 视频生成管道"""

    def __init__(self, ckpt_dir: str, **model_args):  # 🔥 添加ckpt_dir参数
        self.device_type = "cpu"
        super().__init__(ckpt_dir, **model_args)

    def _get_backend(self) -> str:
        return "gloo"

    def _load_model(self):
        logger.info(f"Rank {self.rank}: Loading model on CPU")
        # CPU模型加载逻辑
        # 这里添加你的CPU模型加载代码
        return None  # 临时返回None

    def _generate_video_device_specific(self, request, img, progress_callback=None):
        logger.info(f"Rank {self.rank}: Generating video on CPU")
        # CPU视频生成逻辑
        # 这里添加你的CPU推理代码
        return None  # 临时返回None

    def _save_video(self, video_tensor, output_path: str):
        logger.info(f"Saving video to {output_path}")
        # 创建占位文件
        os.makedirs(os.path.dirname(output_path), exist_ok=True)
        with open(output_path, "wb") as f:
            f.write(b"FAKE_VIDEO_DATA")

    def _log_memory_usage(self):
        logger.info(f"Rank {self.rank} CPU Memory usage tracking not implemented")

    def _empty_cache(self):
        # CPU不需要清空缓存
        pass